% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_oligo.r
\name{count_oligo}
\alias{count_oligo}
\title{Count oligonucleotide frequency}
\usage{
count_oligo(oligo, sequence)
}
\arguments{
\item{oligo}{character (or character vector) of oligonucleotide.}

\item{sequence}{character (or character vector) of DNA (or RNA)
sequence.}
}
\value{
vector of occurences
}
\description{
Counts occurence of one or more oligonucleotides in on or more sequences.
This is wraper around \code{stringr} function
\code{\link[stringr]{str_count}}. Do to its limitations, you cannot search
multiple oligos in multiple sequences at once.
}

