% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fake_island.r
\name{create_fake_island}
\alias{create_fake_island}
\title{Create fake islands}
\usage{
create_fake_island(target, source, islands)
}
\arguments{
\item{target}{sequence that will be modified}

\item{source}{sequence from which islands are taken}

\item{islands}{list of vectors of three elements. First is start position of
island in source sequence, second is start position of island in target
sequence. Third is length of island.}
}
\value{
modified target sequence with short segments from source sequence
}
\description{
Function substitute part of target sequence for part of source sequence,
creating an artificial genomic island.
}

