% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_oligo.r
\name{generate_oligo}
\alias{generate_oligo}
\title{Generating oligonucleotides}
\usage{
generate_oligo(length)
}
\arguments{
\item{length}{of generated oligonucleotides}
}
\value{
character vector of oligonucleotides
}
\description{
Generate every possible oligonucleotide of certain length.
}

