% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credibility_interval.r
\name{get_credibility_interval}
\alias{get_credibility_interval}
\title{Computes credibility interval}
\usage{
get_credibility_interval(density, alpha = c(0.05, 0.025, 0.01))
}
\arguments{
\item{density}{estimated density with \code{\link[stats]{density}} function}

\item{alpha}{vector of significance levels for which credibility interval
will be computed}
}
\value{
list of credibility intervals
}
\description{
Function computes credibility interval for various alphas estimated density.
}

