% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.r
\name{get_signature}
\alias{get_signature}
\title{Get signature}
\usage{
get_signature(sequence, oligos, window = 5000, step = 1000, file = NULL)
}
\arguments{
\item{sequence}{string of DNA bases (ACGT)}

\item{oligos}{character (or character vector) of oligonucleotide.}

\item{window}{size of windows on which frequency is calculated}

\item{step}{amount of bases by which each window is moved #TODO nonsense, better description}

\item{file}{If specified, output will be piped file (or connection) instead}
}
\value{
matrix of frequencies of \code{oligos} with quality of sequence
}
\description{
Get signature for sequence. Signature is frequency of oligonucleotides
computed for smaller overlapping subsets -- windows.
Function \code{get_signature} will partition sequence into individual
windows, neighbouring windows overlap by \code{window-step}.
For these windows, frequency of oligonucleotides is computed as
count of oligonucleotide divided by maximum number of oligonucleotides of
given length in window.
}
\details{
Quality of sequence is also outputted, this is computed as frequency of
non-ACGT characters.
}

