% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_density.r
\name{global_density}
\alias{global_density}
\title{Calculate DIAS using global density method}
\usage{
global_density(signature, oligos = NULL, alpha = c(0.05, 0.025, 0.01))
}
\arguments{
\item{signature}{signature from \code{\link{get_signature}} function}

\item{oligos}{either numeric vector or character vector of oligonucleotides
that will be used. If not specified, whole signature is used.}

\item{alpha}{significance level of scoring intervals.}
}
\value{
DIAS score for given signature.
}
\description{
Function calculate DIAS score from signature using global density method.
}
\details{
First, signature is filtered and \code{quality} is removed. If \code{oligos}
were specified, only specified oligos (either columns, if \code{oligos} was
numeric vector, or specific oligonucleotides, if \code{oligos} was character
vector) are kept. Then global density for all retained oligonucleotide is
calculated and each window of signature is scored if it falls to confidence
interval defined by alpha. Then all DIAS from all oligonucleotides are
summed and final DIAS is returned.
}

