% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.r
\name{hello_world}
\alias{hello_world}
\title{Hello world}
\usage{
hello_world()
}
\description{
Simple hello world function to test basic functionality.
}
\examples{
hello_world()
}
\keyword{hello,}
\keyword{world}

